norm

+++ libft/ft_substr.c
@@ -5,20 +5,20 @@
 /*                                                    +:+ +:+         +:+     */
 /*   By: enja <enja@student.42.fr>                  +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2021/11/17 19:58:05 by yer-retb          #+#    #+#             */
-/*   Updated: 2022/09/17 01:25:40 by enja             ###   ########.fr       */
+/*   Created: 2021/11/23 15:25:08 by enja              #+#    #+#             */
+/*   Updated: 2022/09/18 22:27:18 by enja             ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-char	*ft_substr(const char *str, unsigned int start, size_t len)
+char	*ft_substr(const char *str, int start, int len)
 {
-	char				*s;
-	char				*p;
-	unsigned int		i;
-	unsigned int		j;
-	size_t				s_len;
+	char			*s;
+	char			*p;
+	int				i;
+	int				j;
+	int				s_len;
 
 	s = (char *)str;
 	i = 0;
